feat: implement OpenAPI documentation (#162)
