+++ src/server/run/hpo_genes.rs
@@ -12,7 +12,7 @@ use hpo::{
     Ontology,
 };
 
-use crate::server::WebServerData;
+use crate::server::run::WebServerData;
 
 use super::{CustomError, Match, ResultHpoTerm};
 
@@ -195,12 +195,14 @@ mod test {
         let hpo_doc = fastobo::from_file("tests/data/hpo/hp.obo")?;
         let app = actix_web::test::init_service(
             actix_web::App::new()
-                .app_data(actix_web::web::Data::new(crate::server::WebServerData {
-                    ontology,
-                    ncbi_to_hgnc,
-                    hgnc_to_ncbi,
-                    full_text_index: crate::index::Index::new(hpo_doc)?,
-                }))
+                .app_data(actix_web::web::Data::new(
+                    crate::server::run::WebServerData {
+                        ontology,
+                        ncbi_to_hgnc,
+                        hgnc_to_ncbi,
+                        full_text_index: crate::index::Index::new(hpo_doc)?,
+                    },
+                ))
                 .service(super::handle),
         )
         .await;
