+++ src/server/run/hpo_terms.rs
@@ -9,7 +9,7 @@ use actix_web::{
 };
 use hpo::{annotations::AnnotationId, HpoTerm, HpoTermId, Ontology};
 
-use crate::server::WebServerData;
+use crate::server::run::WebServerData;
 
 use super::{CustomError, ResultGene};
 
@@ -340,7 +340,7 @@ mod test {
         let hpo_doc = fastobo::from_file("tests/data/hpo/hp.obo")?;
         let app = actix_web::test::init_service(
             actix_web::App::new()
-                .app_data(actix_web::web::Data::new(crate::server::WebServerData {
+                .app_data(actix_web::web::Data::new(crate::server::run::WebServerData {
                     ontology,
                     ncbi_to_hgnc,
                     hgnc_to_ncbi,
